PostCommentApiDto Refactor
