+++ src/main/java/kr/or/connect/reservation/dto/api/PostCommentApiDTO.java
@@ -1,4 +1,4 @@
-package kr.or.connect.reservation.dto;
+package kr.or.connect.reservation.dto.api;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class PostCommentDTO {
+public class PostCommentApiDTO {
     String result;
     int productId;
 }
